Write a C program that prints the size of various types on the computer it is compiled and run on.

+++ 0x00-hello_world/6-size.c
@@ -12,7 +12,7 @@ int main(void)
 	printf("Size of a int: %d byte(s)\n", sizeof(int));
 	printf("Size of a long int: %ld byte(s)\n", sizeof(int));
 	printf("Size of a long long int: %lld byte(s)\n", sizeof(int));
-	printf("Size of a float: %f byte(s)\n", sizeof(double));
+	printf("Size of a float: %f byte(s)\n", sizeof(float));
 	return (0);
 }
